+++ versions/1.21.5/src/main/java/com/hypherionmc/craterlib/mixin/events/WhitelistMixin.java
@@ -21,7 +21,7 @@ public abstract class WhitelistMixin<K, V extends StoredUserEntry<K>> {
     @Shadow protected abstract boolean contains(K k0);
 
     @Inject(method = "add", at = @At("HEAD"))
-    private void injectAddEvent(V arg, CallbackInfoReturnable<Boolean> cir) {
+    private void injectAddEvent(V arg, CallbackInfo ci) {
         try {
             if (arg instanceof UserWhiteListEntry entry) {
                 StoredUserEntryAccessor entryAccessor = (StoredUserEntryAccessor) entry;
@@ -32,8 +32,8 @@ public abstract class WhitelistMixin<K, V extends StoredUserEntry<K>> {
         } catch (Exception ignored) {}
     }
 
-    @Inject(method = "remove(Lnet/minecraft/server/players/StoredUserEntry;)Z", at = @At("HEAD"))
-    private void injectRemoveEvent(StoredUserEntry<K> arg, CallbackInfoReturnable<Boolean> cir) {
+    @Inject(method = "remove(Lnet/minecraft/server/players/StoredUserEntry;)V", at = @At("HEAD"))
+    private void injectRemoveEvent(StoredUserEntry<K> arg, CallbackInfo ci) {
         try {
             if (arg instanceof UserWhiteListEntry entry) {
                 StoredUserEntryAccessor entryAccessor = (StoredUserEntryAccessor) entry;